arrumando listagem de produtos, dutosdutos, pagamento fake e cadastro de pedidos

+++ app/src/main/java/com/example/apptcc/model/ItemPedido.java
@@ -6,7 +6,7 @@ public class ItemPedido {
 
     private Produto produto;
     private int quantidade = 0;
-    private BigDecimal valor;
+    private Double valor;
 
     public ItemPedido() {
     }
@@ -27,11 +27,11 @@ public class ItemPedido {
         this.quantidade = quantidade;
     }
 
-    public BigDecimal getValor() {
+    public Double getValor() {
         return valor;
     }
 
-    public void setValor(BigDecimal valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
